+++ actix-web-lab/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased - 2022-xx-xx
 
+- Add `UrlEncodedForm` extractor with const generic size limit that supports collecting multi-value items into a `Vec`.
 
 ## 0.18.6 - 2022-12-08